BSeparate header; buttons for user and langauge
